Refactored plot_sp_predicts for plot type
In preparation for adding observed data to the forecast plot, I added an argument to the function to specify whether the user wants just the forecast (if observed data not aavilable yet) or both forecast + observed. For reasons beyond my comprehension, this required splitting the function into two functions - a data processing function and a plotting function. Otherwise plotting did not work. plot_sp_predicts calls the plotting function, so only one function is still called in the main code to generate the plot.