Refactor forecast (#41)
* refactor PortalForecats.R to run automatically

Before some minor things had to be edited manually to re-run this script every month. This now runs automatically to make a forecast 12 months out from the most recent sample. There is also a lot of changes for clarification of data cleaning and model building. 

* Large refactor of the initial data setup

- Add in more comments
- Use more descriptive variables
- Automatically extract the most recent samples NewMoonNumber to use as a starting point for the next forecast. This is better than hard coding it in every month.
- Create a 6 month lag between rodent observations and weather using a join instead of just subsetting them the same way. Using a join this way is more robust to missing months or 2 trapping sessions in a month.
- Nothing was changed in the actual models except to match some new variable names. Though they will change to some extent because of the new way the lag is calculated.

* account for very rare species

Some species have 0 occurances since the mid 90's, which makes tsgml() balk. Account for that by forecasting 0 abundance for them.

* account for very rare species

Some species have 0 occurances since the mid 90's, which makes tsgml() balk. Account for that by niavely forecasting 0 abundance for them.

* iterate thru species by name instead of index number

The for loop before used the index to pull the abundance for each species, thus depended on the specific ordering of columns. This is more explicit and easer to debug.

* correct species name in results and add total

* fix environmental covariate selection

-use names of env variables instead of list indexes
-more explicit variable names in model comparisons
-account for error prone models, likely due to zero inflation
-fix AIC model selection.
Before with the Poison Env model selection via AIC, the best model AIC was not being updated when a better model was found (see old line 177), so the best model was not necessarily the one being used in the end.

* allow for hindcasted dates

If run normally this will set the Date of a forecast to todays date. But if run with the hindcast option it will set the date to the most recent sample + 1 day so that hindcasts have appropriate forecast dates in the past.

* change weather data.frame to weather_data

otherwise it conflicts with PortalDataSummaries::weather()

* remove rare error causing species