Automatically archive to Zenodo
This change automatically adds a new tag using the current date
and then uses the API to create a release. Since the repo is set up
to archive to Zenodo on release.

This creates a new archive of the entire code base including all of
the predictions. This isn't ideal since there's lots of other stuff
in the repo, but for the time being this is the thing I can get
working.