Remove invalid forecasts
Two monthly forecasts were made that were in error due to a data manipulation
issue resulting from breaking changes to the external `portalr` package and the
underlying Portal Data. The fix for this issue was implemented by @skmorgane in:

https://github.com/weecology/portalr/commit/19ebc06a6a77145dd6f3758a35b7d6a96f5eba23

This issue caused forecasts between 2017-02-22 and 2017-03-24 to be based on
roughly 10x more rodents than were actually present due to a bad join.

Co-authored-by: Shawn Taylor <shawntaylor@weecology.org>
Co-authored-by: Glenda Yenni <glenda@weecology.org>