running the Feb 12 forecast locally
to add to the predictions folder
the Feb 8 forecast had the buggy portalcasting that predicted All twice rather than All and Controls. This forecast should be pretty similar/nearly identical to the Feb 8 one for All but has the proper Controls (and thus proper ensemble), but we'll leave that one in as well for posterity.